// src/shared/enums/settings.ts
export enum settings {
  azureAdClientId = "AZURE_AD_CLIENT_ID",
  azureAdClientSecret = "AZURE_AD_CLIENT_SECRET",
  azureAdGraphQueryUsers = "AZURE_AD_GRAPH_QUERY_USERS",
  azureAdPermissions = "AZURE_AD_PERMISSIONS",
  azureAdTenantId = "AZURE_AD_TENANT_ID",
  azureAutomationWebhookCreateLocalAdUser = "AZURE_AUTOMATION_WEBHOOK_CREATE_LOCAL_AD_USER",
  azureAutomationWebhookUpdateLocalAdUser = "AZURE_AUTOMATION_WEBHOOK_UPDATE_LOCAL_AD_USER",
  cacheTimeoutInSeconds = "CACHE_TIMEOUT_IN_SECONDS",
  cacheTimeoutMultiplierForPhoto = "CACHE_TIMEOUT_MILTIPLIER_FOR_PHOTO",
  cacheUseDatabase = "CACHE_USE_DATABASE",
  disableNewAccountWhenCreated = "DISABLE_NEW_ACCOUNT_WHEN_CREATED",
  excludeUsers = "EXCLUDE_USERS",
  googleClientId = "GOOGLE_CLIENT_ID",
  googleClientSecret = "GOOGLE_CLIENT_SECRET",
  nextAuthSecret = "AUTH_SECRET",
  renderUserPhotoOnClient = "RENDER_USER_PHOTO_ON_CLIENT",
  renderUserPhotoOnServer = "RENDER_USER_PHOTO_ON_SERVER",
  sharepointTenant = "SHAREPOINT_TENANT",
  sharepointSiteUrlHr = "SHAREPOINT_SITE_URL_HR",
  sharepointListTitleEmployees = "SHAREPOINT_LIST_TITLE_EMPLOYEES",
  sharepointListTitleEmployeeDocuments = "SHAREPOINT_LIST_TITLE_EMPLOYEE_DOCUMENTS",
  sharepointFolderNameEmployeeDocuments = "SHAREPOINT_FOLDER_NAME_EMPLOYEE_DOCUMENTS",
  shouldUseMockupDataWhenAuthenticated = "USE_MOCKUP_DATA_WHEN_AUTHENTICATED",
  shouldUseMockupDataWhenUnauthenticated = "USE_MOCKUP_DATA_WHEN_UNAUTHENTICATED",
  shouldUseMockupDataWhenAuthenticatedWithGoogle = "USE_MOCKUP_DATA_WHEN_AUTHENTICATED_WITH_GOOGLE",
  testCreate = "TEST_CREATE",
  testCreateError = "TEST_CREATE_ERROR",
  treeViewCollapseByDefault = "TREEVIEW_COLLAPSE_BY_DEFAULT",
  treeViewCollapseOnRoot = "TREEVIEW_COLLAPSE_ON_ROOT",
  userCardAttributes = "USER_CARD_ATTRIBUTES",
}